Turned wikipedia.py into a behemoth. It seems to be working well now though.
It slices and dices the page, and only keeps relatively pertinent headings.
Throws away cross-references within the text, that style may be suitable for
wikipedia, but probably not for us. It also truncates long lists of things
(bibliographies being the main offenders).

It throws an exception if it's worried it may have the wrong person
(no sign they're a journalist).

Linking it to the database more sensibly is still TODO.



git-svn-id: https://journa-list.googlecode.com/svn/trunk@319 dbcfc70c-e32f-0410-8862-5d2c5c102a8f