Break article needs_indexing out to separate table
New table - article_needs_indexing - which just lists articles which
are queued for indexing in the xapian db.
Less cumbersome than having the needs_indexing flag as a column in the
article table. Because of the way postgresql works, setting
needs_indexing on all articles would cause a rebuild of the entire
article table. Which is slow (ie hours and hours and hours).